+++ pynestml/codegeneration/autodoc_code_generator.py
@@ -19,7 +19,7 @@
 # You should have received a copy of the GNU General Public License
 # along with NEST.  If not, see <http://www.gnu.org/licenses/>.
 
-from typing import List, Sequence, Union
+from typing import Sequence, Union
 
 import datetime
 import os
